LangChain Prompts/requirements.txt
# Langchain Core
langchain-Core
langchain<1.0.0

# Google Gemini (PaLM)- Integration
langchain-google-genai
google-generativeai
google-genai

# Environment Varibale Management
python-dotenv

# Machine Learning Utilities
numpy
scikit-learn


## To run all these libraries in a single go, run this in your terminal: pip install -r requirements.txt